--- .github/workflows/embed-binaries.yaml
@@ -1,19 +1,0 @@
-name: Embed Binaries
-
-on:
-  schedule:
-    - cron: "59 23 * * *"
-  workflow_dispatch:
-
-permissions:
-  issues: write
-
-jobs:
-  embed-binaries:
-    uses: devantler-tech/workflows/.github/workflows/dotnet-embed-binaries.yaml@main
-    secrets: inherit
-    with:
-      BINARY_NAME: kubeconform
-      DOTNET_PROJECT_NAME: Devantler.KubeconformCLI
-      SCRIPT_PATH: scripts/get-kubeconform-binary.sh
-      VERSION_ARGS: -v
