--- Devantler.KubeconformCLI.slnx
@@ -1,13 +1,0 @@
-<Solution>
-  <Configurations>
-    <Platform Name="Any CPU" />
-    <Platform Name="x64" />
-    <Platform Name="x86" />
-  </Configurations>
-  <Folder Name="/src/">
-    <Project Path="src/Devantler.KubeconformCLI/Devantler.KubeconformCLI.csproj" />
-  </Folder>
-  <Folder Name="/tests/">
-    <Project Path="tests/Devantler.KubeconformCLI.Tests/Devantler.KubeconformCLI.Tests.csproj" />
-  </Folder>
-</Solution>
