+++ README.md
@@ -4,48 +4,21 @@
 [![Test](https://github.com/devantler-tech/dotnet-kubeconform-cli/actions/workflows/test.yaml/badge.svg)](https://github.com/devantler-tech/dotnet-kubeconform-cli/actions/workflows/test.yaml)
 [![codecov](https://codecov.io/gh/devantler-tech/dotnet-kubeconform-cli/graph/badge.svg?token=RhQPb4fE7z)](https://codecov.io/gh/devantler-tech/dotnet-kubeconform-cli)
 
-<details>
-  <summary>Show/hide folder structure</summary>
-
-<!-- readme-tree start -->
-```
-.
-├── .github
-│   └── workflows
-├── scripts
-├── src
-│   └── Devantler.KubeconformCLI
-│       └── runtimes
-│           ├── linux-arm64
-│           │   └── native
-│           ├── linux-x64
-│           │   └── native
-│           ├── osx-arm64
-│           │   └── native
-│           ├── osx-x64
-│           │   └── native
-│           ├── win-arm64
-│           │   └── native
-│           └── win-x64
-│               └── native
-└── tests
-    └── Devantler.KubeconformCLI.Tests
-        └── KubeconformTests
-
-22 directories
-```
-<!-- readme-tree end -->
-
-</details>
-
 A simple .NET library that embeds the Kubeconform CLI.
 
 ## 🚀 Getting Started
 
+### Prerequisites
+
+- .NET 9.0 or later
+- [Kubeconform CLI](https://github.com/yannh/kubeconform?tab=readme-ov-file#installation) installed and available in your system's PATH
+
+### Installation
+
 To get started, you can install the package from NuGet.
 
 ```bash
-dotnet add package Devantler.KubeconformCLI
+dotnet add package DevantlerTech.KubeconformCLI
 ```
 
 ## 📝 Usage
@@ -53,7 +26,7 @@ dotnet add package Devantler.KubeconformCLI
 You can execute the Kubeconform CLI commands using the `Kubeconform` class.
 
 ```csharp
-using Devantler.KubeconformCLI;
+using DevantlerTech.KubeconformCLI;
 
 var (exitCode, output) = await Kubeconform.RunAsync(["arg1", "arg2"]);
 ```
